Creación de nuevo componente y fusión de código
Se ha creado un nuevo componente el cual contiene los gráficos y las tablas
By: Antonio Parra, José Moya
